+++ src/visualization/renderers/ascii/index.js
@@ -1,0 +1,10 @@
+/**
+ * ASCII renderer exports
+ */
+
+export { colors, default as colorsDefault } from './colors.js';
+export { createBox } from './box.js';
+export { createTable } from './table.js';
+export { progressBar } from './progress.js';
+export { renderInfoView } from './info.js';
+export { renderCheckView } from './check.js';